requirements.txt
Flask
Flask_script
Flask-Classy
Flask_SQLAlchemy
pymysql
faker
gunicorn
redis